Move RefDescriptions into git package
