+++ enterprise/internal/codeintel/gitserver/client.go
@@ -212,11 +212,6 @@ func ParseCommitGraph(lines []string) *CommitGraph {
 	}
 }
 
-var refPrefixes = map[string]gitdomain.RefType{
-	"refs/heads/": gitdomain.RefTypeBranch,
-	"refs/tags/":  gitdomain.RefTypeTag,
-}
-
 // RefDescriptions returns a map from commits to descriptions of the tip of each
 // branch and tag of the given repository.
 func (c *Client) RefDescriptions(ctx context.Context, repositoryID int) (_ map[string][]gitdomain.RefDescription, err error) {
@@ -225,70 +220,12 @@ func (c *Client) RefDescriptions(ctx context.Context, repositoryID int) (_ map[s
 	}})
 	defer endObservation(1, observation.Args{})
 
-	args := []string{"for-each-ref", "--format=%(objectname):%(refname):%(HEAD):%(creatordate:iso8601-strict)"}
-	for prefix := range refPrefixes {
-		args = append(args, prefix)
-	}
-
-	out, err := c.execGitCommand(ctx, repositoryID, args...)
+	repo, err := c.repositoryIDToRepo(ctx, repositoryID)
 	if err != nil {
 		return nil, err
 	}
 
-	return parseRefDescriptions(strings.Split(out, "\n"))
-}
-
-// parseRefDescriptions converts the output of the for-each-ref command in the RefDescriptions
-// method to a map from commits to RefDescription objects. Each line should conform to the format
-// string `%(objectname):%(refname):%(HEAD):%(creatordate)`, where
-//
-// - %(objectname) is the 40-character revhash
-// - %(refname) is the name of the tag or branch (prefixed with refs/heads/ or ref/tags/)
-// - %(HEAD) is `*` if the branch is the default branch (and whitesace otherwise)
-// - %(creatordate) is the ISO-formatted date the object was created
-func parseRefDescriptions(lines []string) (map[string][]gitdomain.RefDescription, error) {
-	refDescriptions := make(map[string][]gitdomain.RefDescription, len(lines))
-	for _, line := range lines {
-		line = strings.TrimSpace(line)
-		if line == "" {
-			continue
-		}
-
-		parts := strings.SplitN(line, ":", 4)
-		if len(parts) != 4 {
-			return nil, errors.Errorf(`unexpected output from git for-each-ref "%s"`, line)
-		}
-
-		commit := parts[0]
-		isDefaultBranch := parts[2] == "*"
-
-		var name string
-		var refType gitdomain.RefType
-		for prefix, typ := range refPrefixes {
-			if strings.HasPrefix(parts[1], prefix) {
-				name = parts[1][len(prefix):]
-				refType = typ
-				break
-			}
-		}
-		if refType == gitdomain.RefTypeUnknown {
-			return nil, errors.Errorf(`unexpected output from git for-each-ref "%s"`, line)
-		}
-
-		createdDate, err := time.Parse(time.RFC3339, parts[3])
-		if err != nil {
-			return nil, errors.Errorf(`unexpected output from git for-each-ref (bad date format) "%s"`, line)
-		}
-
-		refDescriptions[commit] = append(refDescriptions[commit], gitdomain.RefDescription{
-			Name:            name,
-			Type:            refType,
-			IsDefaultBranch: isDefaultBranch,
-			CreatedDate:     createdDate,
-		})
-	}
-
-	return refDescriptions, nil
+	return git.RefDescriptions(ctx, repo)
 }
 
 // CommitsUniqueToBranch returns a map from commits that exist on a particular branch in the given repository to
